Update registered Keyword if ButtonText changes
In the current implementation, the Speech Keyword is registered on Start() so if the Button text changes (eg. a toggle button that switches between Pin and UnPin) then the new keyword does not get registered. This change will check for the button text change and update the registered keywords appropriately!